"Refactor actions and router"
