+++ routes/go.mod
@@ -1,0 +1,7 @@
+module routes
+
+go 1.16
+
+require github.com/gorilla/mux v1.8.0
+
+replace modules => ../modules
